refactor: rename styles file and remove unused styles
