+++ src/features/user/MFV2/ContributionsMap/ContributionsMap.module.scss
@@ -1,15 +1,3 @@
-.mapCreditsContainer {
-  position: absolute;
-  right: 48px;
-  bottom: 2px;
-  z-index: 1;
-  cursor: pointer;
-}
-
-.mapCredit {
-  padding: 10px;
-  width: 330px;
-}
 .contributionPopup,
 .registeredTreePopup {
   > div {
@@ -21,9 +9,7 @@
   > div:nth-child(1) {
     display: none;
   }
-}
 
-.contributionPopup {
   > div {
     padding: 0px;
   }
@@ -33,12 +19,6 @@
   padding: 5px 16px;
   font-size: 12px;
   white-space: wrap;
-}
-
-.donation-popup {
-  > div {
-    display: none;
-  }
 }
 
 .donationPopupContainer {
@@ -80,20 +60,11 @@
   flex-direction: column;
 }
 
-.donationContainer {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
 .trees {
   font-weight: 700;
   font-size: 12px;
 }
 
-.seperator {
-  font-weight: 700;
-}
 .countryAndTpo {
   display: inline;
   font-size: 10px;
@@ -108,62 +79,8 @@
 
   .seperator {
     margin: 0px 3px;
+    font-weight: 700;
   }
-}
-
-.singleContributionDate {
-  font-size: 7px;
-  color: rgba(130, 130, 130, 1);
-}
-
-.popupDonateButton {
-  padding: 2px 8px !important;
-  font-size: 10px;
-  font-weight: 700;
-  border-radius: 4px;
-  cursor: pointer;
-}
-
-.listOfContributionsContainer {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  justify-content: center;
-  padding: 5px;
-  background: rgba(33, 150, 83, 0.05);
-  gap: 4px;
-  > .contributionInfoContainer {
-    min-width: 110px;
-    background: #fff;
-    border-radius: 4px;
-    padding: 6px;
-    display: flex;
-    justify-content: space-between;
-    > .trees {
-      font-size: 8px;
-      font-weight: 700;
-      line-height: 7px;
-    }
-    .contributionDate {
-      font-size: 6px;
-      line-height: 7px;
-      color: rgba(79, 79, 79, 1);
-    }
-  }
-}
-
-.totalContribution {
-  min-width: 110px;
-  height: 18px;
-  background: rgba(0, 122, 73, 0.1);
-  color: rgba(0, 122, 73, 1);
-  border-radius: 4px;
-  padding: 7px 14px;
-  display: flex;
-  font-size: 6px;
-  font-weight: 700;
-  align-items: center;
-  justify-content: center;
 }
 
 .projectImageInfoContainer {
@@ -178,19 +95,7 @@
     height: 15px;
   }
 }
-.classificationContainer {
-  display: flex;
-  gap: 3.8px;
-  height: 15px;
-  align-items: center;
-}
-.classificationIcon {
-  width: 10px;
-}
-.classification {
-  font-size: 7px;
-  font-weight: 600;
-}
+
 // styles for registered tree popup
 .registeredTreePopupContainer {
   display: flex;
@@ -217,7 +122,3 @@
   color: rgba(130, 130, 130, 1);
 }
 
-.registeredTreeIconContainer {
-  width: 42px;
-  height: 49px;
-}
